use long double  when calculating _beta, _gamma, AON-AttributedGains/-Objective

+++ mt-kahypar/partition/metrics.h
@@ -67,7 +67,7 @@ template<typename PartitionedHypergraph>
 double compute_double_conductance(const PartitionedHypergraph& phg);
 
 template<typename PartitionedHypergraph>
-double compute_double_aon_hypermodularity(const PartitionedHypergraph& phg);
+long double compute_double_aon_hypermodularity(const PartitionedHypergraph& phg);
 
 }  // namespace metrics
 }  // namespace mt_kahypar
